==> code/30_data_cleaning.ipynb <==
# Data Cleaning
This notebook cleans the census data and merges in several other data sources.
### Preliminaries

import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import os
from simpledbf import Dbf5
import shapefile as sp
import json
#pd.set_option('display.height', 1000)
#pd.set_option('display.max_rows', 1000)
pd.set_option('display.max_columns', 100)
#pd.set_option('display.width', 1000)

with open("data_dir.txt") as f:
    data_dir = f.read()

## Agricultural Census 

[ICPSR 35206](https://www.icpsr.umich.edu/web/ICPSR/studies/35206), collected by Haines, Fishback, and Rhode.

### Trim the sample

# Trim down to only Continental US

data = pd.read_csv(data_dir + 'clean_data/agcensus_20-40.csv')
data.sort_values(['year','county'], inplace=True)
data = data[(data.level!=3)] 
cols = data.columns.tolist()
cols1 = cols[:13] + cols[14:]
cols1.insert(0,cols[13])
df = data[cols1]
df = df.query('name!="ALASKA" and name!="HAWAII" and name!="DIST COLUMBIA"')
df = df.query('FIPS_state!=15 and FIPS_state!=2 and FIPS_state!=11')
#df = df[(df.FIPS!=56047)]
# Drop the index number from the merge
#df= df.drop('Unnamed: 0',axis=1)
df = df.query('county!=9999')

df[(df.level ==2)]['name'].nunique()


df.loc[df.year==1920,'farm_value'] = df.loc[df.year==1920,'farm_value_land'] +\
                                     df.loc[df.year==1920,'farm_value_buildings'] 

df = df.drop("Unnamed: 0",axis=1)

df = df.iloc[:,:6].join(df.iloc[:,6:].apply(pd.to_numeric,errors='coerce',axis=1))

### Crop yields

# Convert tons of corn silage to bushels
# Conversion rate for tons per acre to bushels per acre is from:
# http://cdp.wisc.edu/jenny/crop/estimating.pdf


df['corn_silage_tonsacres_tem']   = df['corn_silage_tons']/df['corn_silage_acres']
df['corn_silage_bushacres_tem']   = df['corn_silage_bushels']/df['corn_silage_acres']
df['corn_silage_bush_acres5_tem'] = df['corn_silage_tonsacres_tem'].fillna(0)*5 + df['corn_silage_bushacres_tem'].fillna(0)
df['corn_silage_bush_acres7_tem'] = df['corn_silage_tonsacres_tem'].fillna(0)*7 + df['corn_silage_bushacres_tem'].fillna(0)
df['corn_grain_bush_acres_tem']   = df['corn_grain_bushels']/df['corn_grain_acres']
df['corn_bush_acres_tem']         = df['corn_bushels']/df['corn_acres']

df['total_corn_bushels5'] = df['corn_bushels'].fillna(0) + df['corn_grain_bushels'].fillna(0) +\
                           df['corn_silage_bushels'].fillna(0) + df['corn_silage_tons'].fillna(0)*5
    
df['total_corn_bushels7'] = df['corn_bushels'].fillna(0) + df['corn_grain_bushels'].fillna(0) +\
                            df['corn_silage_bushels'].fillna(0) + df['corn_silage_tons'].fillna(0)*7
df['total_corn_acres'] =    df['corn_acres'].fillna(0) + df['corn_grain_acres'].fillna(0) + df['corn_other_acres'].fillna(0) + \
                            df['corn_silage_acres'].fillna(0)
df['corn_acres_total_tem'] = df['corn_acres'].fillna(0) + df['corn_grain_acres'].fillna(0) + df['corn_other_acres'].fillna(0) + \
                                df['corn_silage_acres'].fillna(0)
df['corn_acres_subtotal_tem'] = df['corn_acres'].fillna(0) + df['corn_grain_acres'].fillna(0) + \
                                df['corn_silage_acres'].fillna(0)

df['adjustment_tem'] = df['corn_acres_subtotal_tem'].divide(df['corn_acres_total_tem'],fill_value=0)

df['corn_yield_total5']   = (df['corn_bush_acres_tem'].fillna(0) + df['corn_silage_bush_acres5_tem'] + df['corn_grain_bush_acres_tem'].fillna(0))
df['corn_yield_total7']   = (df['corn_bush_acres_tem'].fillna(0) + df['corn_silage_bush_acres7_tem'] + df['corn_grain_bush_acres_tem'].fillna(0))
df['corn_yield5']         = df['adjustment_tem']*df.corn_yield_total5
df['corn_yield7']         = df['adjustment_tem']*df.corn_yield_total7

df = df.replace(np.inf, 0)

tempsc = list(df.filter(regex="corn").filter(regex="tem").columns)
df.drop(tempsc, axis=1, inplace=True)
df.filter(regex="corn").mean()

# Find the yield on wheat by totalling up

df['total_wheat_acres']     = df['wheat_acres'].fillna(0) + df['wheat_grain_acres'].fillna(0) + df['wheat_emmersp_acres'].fillna(0) + df['wheat_thresh_acres'].fillna(0) +df['wheat_threshed_acres'].fillna(0)
df['total_wheat_bushels']   = df['wheat_bushels'].fillna(0) + df['wheat_durummac_bushels'].fillna(0) + df['wheat_emmersp_bushels'].fillna(0) + df['wheat_spring_bushels'].fillna(0) +df['wheat_winter_bushels'].fillna(0) 

df['wheat_yield']     = df['total_wheat_bushels'].divide(df['total_wheat_acres'],fill_value=0)
df = df.replace(np.inf, 0)

tempsw = list(df.filter(regex="wheat").filter(regex="tem").columns)
df.drop(tempsw, axis=1, inplace=True)

# Above code puts an NA when there should be a zero, this code puts a zero if zero acres are reported.

miss = df[(np.isnan(df.wheat_yield)==True)].filter(regex="wheat")
df["boolsum"] = df.filter(regex="wheat").notnull().sum(axis=1)
df.loc[np.isnan(df.wheat_yield)&df.boolsum>0,'wheat_yield'] = 0

df.drop(["boolsum"], axis=1, inplace=True)
df.filter(regex="wheat").mean()

# Find the bean yield
beanacre = [x for x in list(df.filter(regex='bean').filter(regex='acre')) if 'soy' not in x and 'bool' not in x and 'total' not in x ]

df['total_bean_acres'] = 0
df['bean_acre_bool'] = 0
for var in beanacre:
    df['total_bean_acres'] = df['total_bean_acres'] +df[var].fillna(0)
    df['bean_acre_bool'] =( ~np.isnan(df[var])).astype(int) + df['bean_acre_bool']
df.loc[df.bean_acre_bool==0,'total_bean_acres'] = np.nan

beanbush = [x for x in list(df.filter(regex='bean').filter(regex='bushel')) if 'soy' not in x ]
df['total_bean_bshls'] = 0
df['bean_bush_bool'] = 0
for var in beanbush:
    df['total_bean_bshls'] = df['total_bean_bshls'] +df[var].fillna(0)
    df['bean_bush_bool'] =( ~np.isnan(df[var])).astype(int) + df['bean_bush_bool']
df.loc[df.bean_bush_bool==0,'total_bean_bshls'] = np.nan

soybush = [x for x in list(df.filter(regex='soybean').filter(regex='bushel')) if 'bool' not in x and 'total' not in x  ]
df['total_soy_bshls'] = 0
df['soy_bush_bool'] = 0
for var in soybush:
    df['total_soy_bshls'] = df['total_soy_bshls'] +df[var].fillna(0)
    df['soy_bush_bool'] =( ~np.isnan(df[var])).astype(int) + df['soy_bush_bool']
df.loc[df.soy_bush_bool==0,'total_soy_bshls'] = np.nan

soyacres = [x for x in list(df.filter(regex='soybean').filter(regex='acres')) if 'bool' not in x and 'total' not in x  ]
df['total_soy_acres'] = 0
df['soy_acres_bool'] = 0
for var in soyacres:
    df['total_soy_acres'] = df['total_soy_acres'] +df[var].fillna(0)
    df['soy_acres_bool'] =( ~np.isnan(df[var])).astype(int) + df['soy_acres_bool']
df.loc[df.soy_acres_bool==0,'total_soy_acres'] = np.nan

df['soybean_yield'] = df['total_soy_bshls']/df['total_soy_acres'].replace(np.inf,np.nan) 

df.drop(list(df.filter(regex='bool').columns),axis=1,inplace=True)
df.loc[(df.total_soy_acres==0)|(df.total_soy_bshls==0),'soybean_yield'] = 0

#### Animal product value

df['animal_rev_1920'] = df['dairy_value'] + df['chicken_egg_value'] + df['wool_value']
df['animal_revenue'] = df['dairy_value'] + df['chicken_value'] + df['egg_value'] + df['wool_value']
df.loc[df.year==1920,'animal_revenue'] = df['animal_rev_1920']

## Spatial Data (created in QGIS)
### - Climate data
[Prism Climate Group](https://prism.oregonstate.edu/), publishes average temperature and rainfall.
### - Soil measurements
[GAEZ Soil Suitability for Corn and Wheat](http://www.fao.org/nr/gaez/en/), published by the FAO.

df['FIPS']=[str(x).split(".")[0].zfill(5) for x in df.FIPS]

# Read in the spatial dataset and climate data
dbf_precip   = Dbf5(data_dir + 'spatial_data/ppt_total1.dbf',codec='utf-8')
dbf_tmean    = Dbf5(data_dir + 'spatial_data/tmean_total.dbf',codec='utf-8')

df_ppt = dbf_precip.to_dataframe()
df_tmn = dbf_tmean.to_dataframe()

# Two extinct Georgia counties have to be taken out
df_tmn = df_tmn.loc[~((df_tmn.ICPSRNAM=='MILTON')&(df_tmn.STATENAM=='Georgia'))]
df_tmn = df_tmn.loc[~((df_tmn.ICPSRNAM=='CAMPBELL')&(df_tmn.STATENAM=='Georgia'))]

df_ppt = df_ppt.loc[~((df_ppt.ICPSRNAM=='MILTON')&(df_ppt.STATENAM=='Georgia'))]
df_ppt = df_ppt.loc[~((df_ppt.ICPSRNAM=='CAMPBELL')&(df_ppt.STATENAM=='Georgia'))]

df_ppt = df_ppt[['_ppt_mean','_ppt_std','_ppt_var','FIPS','year','ICPSRNAM']]
df_ppt.columns = ['rain_mean','rain_std','rain_var','FIPS','year','ICPSRNAM']

df_tmn = df_tmn[['_tmn_mean','_tmn_std','_tmn_var','FIPS','year']]
df_tmn.columns = ['temp_mean','temp_std','temp_var','FIPS','year']

df_ppt_tmn = df_ppt.merge(df_tmn)


# Read in soils
soils = Dbf5(data_dir + 'spatial_data/soils.dbf', codec='utf-8')\
            .to_dataframe().drop(['wheat_mean','wheat_sum','wheat_coun','wcount','wsum','wmean',\
                                  'corncount','cornsum','wheatcount','wheatsum'],axis=1)
soils['FIPS'] = soils.countyid

# Two extinct Georgia counties have to be taken out
soils = soils.loc[~((soils.ICPSRNAM=='MILTON')&(soils.ICPSRST=='44'))]
soils = soils.loc[~((soils.ICPSRNAM=='CAMPBELL')&(soils.ICPSRST=='44'))]

df_pts = df_ppt_tmn.merge(soils[['FIPS','cornmean','wheatmean','X_CENTROID','Y_CENTROID','ICPSRST','ICPSRCTY']],how='left',on='FIPS')

dfmerge = pd.merge(df,df_pts,on=['FIPS','year'],how='left')

# Some counties misnamed in spatial data.

boolvec = dfmerge['name'] ==dfmerge['ICPSRNAM']
name_i = list(boolvec[boolvec==False].index)
dfmerge[['name','ICPSRNAM','FIPS']].loc[name_i]
name_1 = name_i[:3] +name_i[4:]
dfmerge = dfmerge.drop(dfmerge.index[name_1])

# Some columns were misnamed and must be merged in to each other.

dfmerge['farm_number'] = pd.concat([dfmerge['farm_number'].dropna(), dfmerge['farms_number'].dropna()]).reindex_like(dfmerge)
dfmerge['tenant_number'] = pd.concat([dfmerge['tenant_number'].dropna(), dfmerge['tenants_number'].dropna()]).reindex_like(dfmerge)
dfmerge['mort_farms_no'] = pd.concat([dfmerge['mort_farms_no'].dropna(), dfmerge['mort_farms_num'].dropna()]).reindex_like(dfmerge)

dfmerge = dfmerge.drop(['farms_number','tenants_number','mort_farms_num'],axis=1)

# Only include 1920 counties
FIPS_i = list(dfmerge.FIPS.value_counts()[(dfmerge.FIPS.value_counts()<5)].index)
dfmerge = dfmerge[-dfmerge.FIPS.isin(FIPS_i)]

## PCA information

Produced from the notebook "PCA_calculations"

PCAs = pd.read_csv(data_dir+"clean_data/PCA_info.csv")

dfmerge['FIPS'] = dfmerge.FIPS.astype(int)

dfmerge = dfmerge.merge(PCAs,how='left')

#### PCA District Dummy

# Make district dummies
districts= {1:['9','23','25','33','34','36','44','50'],
            2:['10','24','42','51','54'],
            3:['37','45','13','12'],
            4:['39','18','21','47'],
            5:['1','28','22'],
            6:['17','29','5'],
            7:['38','27','55','26'],
            8:['56','46','31','19'],
            9:['8','35','40','20'],
            10:['48'],
            11:['6','32','49','4'],
            12:['53','41','16','30']}

dfmerge['FIPS_state'] = dfmerge.FIPS_state.astype(int).astype(str)
dfmerge['pca_district'] = np.zeros((len(dfmerge),1))
for i,j in districts.items():
    dfmerge.loc[dfmerge.FIPS_state.isin(set(j)),'pca_district'] = int(i)
    
# Categories to use in analysis
dfmerge['pca_district_groups'] = 'None'
dfmerge.loc[(dfmerge['pca_district']==3)|(dfmerge['pca_district']==5),'pca_district_groups'] = 'South'
dfmerge.loc[dfmerge['pca_district']<3,'pca_district_groups'] = 'Northeast'
dfmerge.loc[(dfmerge['pca_district']==4)|(dfmerge['pca_district']==6),'pca_district_groups'] = 'Midwest'
dfmerge.loc[(dfmerge['pca_district']>6)&(dfmerge['pca_district']<11),'pca_district_groups'] = 'Great Plains'
dfmerge.loc[(dfmerge['pca_district']>10),'pca_district_groups'] = 'West'

## PCA State Level data
Collected from the Farm Credit Administration annual reports

# Merge in the state level data
pcastate = pd.read_csv(data_dir + "clean_data/pca_state_vars.csv")
pcastate['FIPS_state'] = pcastate.FIPS_state.astype(str)
dfmerge = pd.merge(dfmerge, pcastate, how='left', on= ['FIPS_state','year']).sort_values(['year','FIPS'])
dfmerge.iloc[:,-10:]= dfmerge.iloc[:,-10:].fillna(0)
dfmerge.drop('NAME', axis=1, inplace=True)

## Erosion Data
From "The Enduring Impact of the American Dust Bowl: Short- and Long-Run Adjustments to Environmental Catastrophe" by Hornbeck (2012).

https://www.aeaweb.org/articles?id=10.1257/aer.102.4.1477

erosion = pd.read_csv(data_dir + "clean_data/hornbeck12_erosion.csv")

erosion['fips']=[str(x).split(".")[0].zfill(5) for x in erosion.fips]

erosion['fips'] = erosion['fips'].astype(int)

dfmerge = dfmerge.merge(erosion.rename(columns={'fips':'FIPS'}),how='left',on='FIPS')

## Deal with extinct counties

dfmerge= dfmerge.loc[~((dfmerge.ICPSRNAM=='MILTON')&(dfmerge.ICPSRST=='44'))]
dfmerge= dfmerge.loc[~((dfmerge.ICPSRNAM=='CAMPBELL')&(dfmerge.ICPSRST=='44'))]

## New Deal Data

#### Source: http://www.u.arizona.edu/~fishback/Published_Research_Datasets.html

This data is from the paper "Can the New Deal's Three R's Be Rehabilitated?: A Program-by-Program, County-by-County Analysis" by Fishback, Kantor, and Wallis

newdeal = pd.read_csv(data_dir + 'newdeal/new_deal_spend.csv').rename(columns={'STATE':'ICPSRST','NDCODE':'ICPSRCTY'})
newdeal = newdeal[['ICPSRST','ICPSRCTY','FERA','CWA','WPA','PUBASS','PWAF','PWANF1','PWANF2',\
                     'PRA','PBA','AAA','FCA','FSARR2','FSALO','REA','RFC',\
                     'HOLC','INS','USHALC','USHAH','NDEXP','RELIEF','PUBWOR',\
                     'LOAN','MEAN9628','ROOSMMN2','STD9632','PCTVT32','HTEN73B']]
newdeal = newdeal.rename(columns={'NDEXP':'total_grants','RELIEF':'total_relief','PUBWOR':'public_works','LOAN':'total_loans'})
newdeal['ICPSRST'] = newdeal['ICPSRST'].astype(str)
newdeal['ICPSRCTY'] = newdeal['ICPSRCTY'].astype(str)
dfmerge = dfmerge.merge(newdeal, how='left',on=['ICPSRST','ICPSRCTY'])

## Hybrid Corn data

Source: NASS (1945). Agricultural Statistics. Technical report, USDA, Washington, DC. Table 46.

state_names = pd.read_csv(data_dir+"clean_data/state_names.csv")
hybridcorn = pd.read_csv(data_dir + 'clean_data/hybrid_corn.csv')
hybridcorn = hybridcorn.merge(state_names,how='left').drop_duplicates()

dfmerge['FIPS_state'] = dfmerge['FIPS_state'].astype(int)

dfmerge = dfmerge.merge(hybridcorn[['FIPS_state','year','corn_hybrid_pct']],how='left',on=['FIPS_state','year'])

## FDIC Bank Data

Source: ICPSR 7 https://www.icpsr.umich.edu/icpsrweb/ICPSR/studies/7?q=fdic

rawdata = pd.read_csv(data_dir + 'fdic/fdicdata.csv')

with open(data_dir + 'fdic/fdic_var_labels.json') as f:
    tmp = f.read()
labels = json.loads(tmp)

#dictionary for the new variable names
new_names = {"A_":"deps_all","B_":"deps_sus",\
             "C_":"no_banks_all","D_":"no_banks_sus",\
             "E_":"deps_natl","F_":"deps_natl_sus",\
             "G_":"deps_st","H_":"deps_st_sus",\
             "I_":"no_banks_natl","J_":"no_banks_natl_sus",\
             "K_":"no_banks_st","L_":"no_banks_st_sus"}

rawdata = rawdata.rename(columns=labels)

rawdata['index'] = rawdata.index

clean_annual = pd.wide_to_long(rawdata,list(new_names.keys()),i="index",j="year_").rename(columns = new_names)
clean_annual = clean_annual.reset_index().drop(['index','YEAR','CATALOG_ENTRY_NMBR','CONG_DIST_NMBR','TABLE_NUMBER'],axis=1).rename(columns ={"year_":"year","COUNTY_NAME":"name",'DATA TYPE':'level'})

clean_annual = clean_annual.replace(-9,np.nan)
clean_annual = clean_annual.replace(np.inf,np.nan)


clean_annual['ID'] = [str(x).split(".")[0].zfill(4) for x in clean_annual['ID']]

clean_annual['ICPSRST'] = [str(x).zfill(2) for x in clean_annual['ICPSRST'].astype(float).astype(int)]

clean_annual['icpsr_id'] = clean_annual['ICPSRST'] + clean_annual['ID']  

clean_annual.sort_values(['icpsr_id','year'], inplace=True)

clean_annual['no_banks_sus_fd'] = clean_annual.groupby(['icpsr_id'])['no_banks_sus'].transform(lambda x: x.diff())
clean_annual.sort_index(inplace=True)
clean_annual['deps_all_lag1'] = clean_annual.groupby('icpsr_id')['deps_all'].shift()
clean_annual['percent_deps_sus'] = clean_annual['deps_sus'] /clean_annual['deps_all_lag1']
clean_annual['pct_deps_sus'] = clean_annual['deps_sus']/clean_annual['deps_all']

fdic_losses = clean_annual[(clean_annual.year<1933)&(clean_annual.year>1928)]\
                                .groupby('icpsr_id')['pct_deps_sus'].mean().reset_index()


dfmerge['ICPSRCTY_id'] = [str(x).zfill(4) for x in dfmerge['ICPSRCTY']]
dfmerge['ICPSRST_id'] = [str(x).zfill(2) for x in dfmerge['ICPSRST']]

dfmerge['icpsr_id']   = dfmerge['ICPSRST_id'] + dfmerge['ICPSRCTY_id']

dfmerge = dfmerge.drop(['ICPSRCTY_id','ICPSRST_id'],axis=1)

fdic_losses['icpsr_id'] = fdic_losses['icpsr_id'].apply(lambda x: str(x).zfill(5))

dfmerge = dfmerge.merge(clean_annual.drop(['name','level','pct_deps_sus'],axis=1), how='left', on=['icpsr_id','year'])

dfmerge = dfmerge.merge(fdic_losses, how='left', on=['icpsr_id'])

# Write data

dfmerge = dfmerge.merge(state_names,how='left').drop_duplicates()

dfmerge.to_csv(data_dir + 'clean_data/fcs_final.csv',index=False)
